Hotfix bug v1.0.7

See merge request hypersmart/redsun/user-center-provider!9

+++ usercenter-provider/src/main/java/com/hypersmart/usercenter/model/RsunUserStarLevell.java
@@ -31,7 +31,7 @@ public class  RsunUserStarLevell implements Serializable {
 
     @Column(name = "xz_num")
     @ApiModelProperty("个人勋章数")
-    private Double xzNum;
+    private Integer xzNum;
 
     public String getUcUserId() {
         return ucUserId;
@@ -65,11 +65,11 @@ public class  RsunUserStarLevell implements Serializable {
         this.totalCoin = totalCoin;
     }
 
-    public Double getXzNum() {
+    public Integer getXzNum() {
         return xzNum;
     }
 
-    public void setXzNum(Double xzNum) {
+    public void setXzNum(Integer xzNum) {
         this.xzNum = xzNum;
     }
